rework gui in waveform
