+++ jackutils/jackgui.c
@@ -8,6 +8,7 @@
 #include <jackgui.h>
 #include <stdlib.h>
 #include <stdio.h>
+#include <assert.h>
 #include <GLFW/glfw3.h>
 
 static void glfwinit__() {
@@ -21,32 +22,44 @@ static void error_callback(int error, const char* description) {
     fprintf(stderr, "Error: %s\n", description);
 }
 
-#define NK_GLFW_GL2_IMPLEMENTATION
-#include <ext/nuklear_glfw_gl2.h>
+#define NK_GLFW_GL3_IMPLEMENTATION
+#include <ext/nuklear_glfw_gl3.h>
 
 /*
  * Initializes GUI context.
  * @ret NULL in case of error
  */
+
+static GLFWwindow* window;
+
 JG_API jg_ctx_t* jg_init(const char* t, int w, int h) {
-	GLFWwindow* window;
 	jg_ctx_t* ctx;
-	struct nk_font_atlas* atlas;
 
 	glfwSetErrorCallback(error_callback);
 	glfwinit__();
 
-	glfwWindowHint(GLFW_SCALE_TO_MONITOR, GLFW_TRUE);
-
+	glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
+	glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 3);
+	glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_COMPAT_PROFILE);
+#ifdef __APPLE__
+	glfwWindowHint(GLFW_OPENGL_FORWARD_COMPAT, GL_TRUE);
+#endif
 	window = glfwCreateWindow(w, h, t, NULL, NULL);
-	if (!window) {
-		return NULL;
-	}
-
+	assert(window != NULL && "can't create glfw window!");
 	glfwMakeContextCurrent(window);
+
+	int width, height;
+	glfwGetWindowSize(window, &width, &height);
+	glViewport(0, 0, width, height);
+
 	ctx = nk_glfw3_init(window, NK_GLFW3_INSTALL_CALLBACKS);
+	
+	struct nk_font_atlas *atlas;
 	nk_glfw3_font_stash_begin(&atlas);
+    /*struct nk_font *droid = nk_font_atlas_add_from_file(atlas, "../../../extra_font/DroidSans.ttf", 14, 0);*/
 	nk_glfw3_font_stash_end();
+	/*nk_style_set_font(ctx, &droid->handle);*/
+	return ctx;
 }
 
 /*
@@ -66,19 +79,37 @@ JG_API int  jg_begin(jg_ctx_t* ctx) {
 	int width, height;
 	struct nk_rect area;
 	
-	glfwGetWindowSize(glfw.win, &width, &height);
-
+	glfwGetWindowSize(window, &width, &height);
 	area = nk_rect(0.f, 0.f, (float) width, (float) height);
-	nk_window_set_bounds(ctx, "", area);
 
-	glClear(GL_COLOR_BUFFER_BIT);
 	nk_glfw3_new_frame();
 	return nk_begin(ctx, "", area, 0);
 }
 
+#define MAX_VERTEX_BUFFER 512 * 1024
+#define MAX_ELEMENT_BUFFER 128 * 1024
+
 JG_API void jg_end(jg_ctx_t* ctx) {
 	nk_end(ctx);
-	nk_glfw3_render(NK_ANTI_ALIASING_ON);
-	glfwSwapBuffers(glfw.win);
+	
+	int width, height;
+	struct nk_rect area;
+	
+	glfwGetWindowSize(window, &width, &height);
+	glViewport(0, 0, width, height);
+	glClear(GL_COLOR_BUFFER_BIT);
+	glClearColor(0,0,0,1);
+	area = nk_rect(0.f, 0.f, (float) width, (float) height);
+
+	nk_glfw3_render(NK_ANTI_ALIASING_ON, MAX_VERTEX_BUFFER, MAX_ELEMENT_BUFFER);
+	glfwSwapBuffers(window);
 	glfwWaitEventsTimeout(1.0);
+}
+
+JG_API int  jg_should_close(jg_ctx_t*) {
+	return glfwWindowShouldClose(window);
+}
+
+JG_API void jg_request_redraw(jg_ctx_t*) {
+	glfwPostEmptyEvent();
 }
